Fixed animation for the left set of rays in main menu
I assume this was a bug from the original game as there were preset values for the positions of all ray pieces. The center ray group and right ray group have their ray pieces at a position of 0, 0, 0. This may need to be reverted or adjusted in the future but for now it looks correct.